refactor: extract multi-select-combo-box logic into mixins
