+++ packages/multi-select-combo-box/package.json
@@ -37,6 +37,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/a11y-base": "24.6.0-alpha9",
     "@vaadin/combo-box": "24.6.0-alpha9",
